bug causing files to be deleted rather than moved
git-svn-id: http://svn.vuze.com/public/client@5217 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/util/FileUtil.java
@@ -824,6 +824,9 @@ public class FileUtil {
 			return( true );
 
 		}else{
+			
+			boolean		success	= false;
+			
 
 			// can't rename across file systems under Linux - try copy+delete
 
@@ -863,6 +866,8 @@ public class FileUtil {
 					throw( new Exception( "Failed to delete '" + from_file.toString() + "'"));
 				}
 				
+				success	= true;
+				
 				return( true );
 				
 			}catch( Throwable e ){		
@@ -891,9 +896,12 @@ public class FileUtil {
 					}
 				}
 				
-				if ( to_file.exists()){
+				if ( !success ){
 					
-					to_file.delete();
+					if ( to_file.exists()){
+						
+						to_file.delete();
+					}
 				}
 			}
 		}
